+++ source/Application/Model/FileChecker.php
@@ -58,13 +58,6 @@ class FileChecker
     protected $_sVersion = "";
 
     /**
-     * Revision of THIS OXID eShop
-     *
-     * @var string
-     */
-    protected $_sRevision = "";
-
-    /**
      * base directory
      *
      * @var mixed
@@ -143,28 +136,6 @@ class FileChecker
     public function getEdition()
     {
         return $this->_sEdition;
-    }
-
-    /**
-     * Revision setter
-     *
-     * @param string $sRevision Revision
-     */
-    public function setRevision($sRevision)
-    {
-        if (!empty($sRevision)) {
-            $this->_sRevision = $sRevision;
-        }
-    }
-
-    /**
-     * Revision getter
-     *
-     * @return string
-     */
-    public function getRevision()
-    {
-        return $this->_sRevision;
     }
 
     /**
@@ -288,7 +259,7 @@ class FileChecker
         $aParams = [
             'job' => 'existsversion',
             'ver' => $this->getVersion(),
-            'rev' => $this->getRevision(),
+            'rev' => false,
             'edi' => $this->getEdition(),
         ];
 
@@ -308,7 +279,7 @@ class FileChecker
             \OxidEsales\Eshop\Core\Registry::getLang()->translateString('OXDIAG_ERRORMESSAGEVERSIONDOESNOTEXIST'),
             $this->getEdition(),
             $this->getVersion(),
-            $this->getRevision()
+            false
         );
 
         $this->_sErrorMessage .= $sError;
@@ -401,7 +372,7 @@ class FileChecker
         $aParams = [
             'job' => 'md5check',
             'ver' => $this->getVersion(),
-            'rev' => $this->getRevision(),
+            'rev' => false,
             'edi' => $this->getEdition(),
             'fil' => $sFile,
             'md5' => $sMD5,