+++ source/Core/OnlineLicenseCheck.php
@@ -259,8 +259,6 @@ class OnlineLicenseCheck
      *
      * @param array $serials Array of serials to add to request.
      *
-     * @throws oxException
-     *
      * @return \OxidEsales\Eshop\Core\OnlineLicenseCheckRequest
      */
     protected function formRequest($serials)
@@ -270,7 +268,7 @@ class OnlineLicenseCheck
         /** @var \OxidEsales\Eshop\Core\OnlineLicenseCheckRequest $request */
         $request = oxNew(\OxidEsales\Eshop\Core\OnlineLicenseCheckRequest::class);
 
-        $request->revision = $config->getRevision();
+        $request->revision = false;
 
         $request->keys = ['key' => $serials];