+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- source/Application/views/admin/tpl/shop_license.tpl
+
 ### Deprecated
 
 ### Removed
@@ -24,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - \OxidEsales\EshopCommunity\Application\Controller\Admin\NavigationTree::_getDynMenuLang()
 - \OxidEsales\EshopCommunity\Application\Controller\Admin\NavigationTree::_getDynMenuUrl()
 - \OxidEsales\EshopCommunity\Application\Component\Widget::getCompareItemsCnt()
+- \OxidEsales\EshopCommunity\Core\Config::getRevision()
+- \OxidEsales\EshopCommunity\Core\Controller\BaseController::getRevision()
 - template source/Application/views/admin/tpl/dyn_econda.tpl
 - template source/Application/views/admin/tpl/dynscreen.tpl
 - template source/Application/views/admin/tpl/dynscreen_list.tpl